# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AEO Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /runs:
    get:
      operationId: listRuns
      tags: [runs]
      summary: Recent simulation runs for a company
      parameters:
        - name: companyId
          in: query
          required: true
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 25 }
        - name: model
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Recent runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromptRun"
  /runs/{id}:
    get:
      operationId: getRun
      tags: [runs]
      summary: Full run detail including answer text and citations
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptRunDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /sources/domains:
    get:
      operationId: listDomains
      tags: [sources]
      summary: Domains that contribute to AI answers, ranked by retrievals
      parameters:
        - name: companyId
          in: query
          required: true
          schema: { type: integer }
        - name: view
          in: query
          schema:
            type: string
            enum: [top, new, trending, losing]
            default: top
        - name: days
          in: query
          schema: { type: integer, default: 30 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Domains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DomainStat"
  /sources/domain-types:
    get:
      operationId: getDomainTypeDistribution
      tags: [sources]
      summary: Retrieval share by domain type
      parameters:
        - name: companyId
          in: query
          required: true
          schema: { type: integer }
        - name: days
          in: query
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: Distribution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainTypeDistribution"
  /sources/urls:
    get:
      operationId: listUrls
      tags: [sources]
      summary: Individual URLs cited in AI answers
      parameters:
        - name: companyId
          in: query
          required: true
          schema: { type: integer }
        - name: days
          in: query
          schema: { type: integer, default: 30 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UrlStat"
...
components:
...
  schemas:
...
    ErrorMessage:
      type: object
      properties:
        error: { type: string }
      required: [error]
    CitationDiagnostics:
      type: object
      description: Structured citation-extraction diagnostics for a run
      required: [metadataEvents, invalidUrls, unknownAnnotationTypes, unknownShapes, redirectsResolved, redirectsFailed]
      properties:
        extracted: { type: integer }
        metadataEvents: { type: integer }
        invalidUrls: { type: integer }
        unknownAnnotationTypes:
          type: array
          items: { type: string }
        unknownShapes: { type: integer }
        redirectsResolved: { type: integer }
        redirectsFailed: { type: integer }
...
    PromptRun:
      type: object
      required: [id, promptId, promptText, model, brandMentioned, createdAt, citationCount]
      properties:
        id: { type: integer }
        promptId: { type: integer }
        promptText: { type: string }
        model: { type: string }
        brandMentioned: { type: boolean }
        brandPosition: { type: ["integer", "null"] }
        createdAt: { type: string }
        citationCount: { type: integer }
        answerPreview: { type: ["string", "null"] }
        searchStatus:
          type: ["string", "null"]
          enum:
            [
              provider_cited,
              search_no_citations,
              extraction_failed,
              provider_search,
              unsupported,
              tool_rejected,
              null,
            ]
          description: "provider_cited = search + verified citations; search_no_citations = search performed, provider returned no citation metadata; extraction_failed = metadata present but unverifiable; provider_search = legacy search label (pre-outcome-model); null = legacy run (search behavior unknown)"
        citationEligible: { type: ["boolean", "null"] }
    PromptRunDetail:
      allOf:
        - $ref: "#/components/schemas/PromptRun"
        - type: object
          required: [answerText, citations, citationDiagnostics]
          properties:
            answerText: { type: string }
            citations:
              type: array
              items:
                $ref: "#/components/schemas/Citation"
            citationDiagnostics:
              description: "Stored citation-extraction diagnostics for this run; null for legacy runs that predate diagnostics capture"
              oneOf:
                - $ref: "#/components/schemas/CitationDiagnostics"
                - type: "null"
...
    Citation:
      type: object
      required: [id, domain, domainType, position]
      properties:
        id: { type: integer }
        domain: { type: string }
        url: { type: ["string", "null"] }
        domainType: { type: string }
        position: { type: integer }
        provenance: { type: ["string", "null"], description: "'provider' = verified provider-returned metadata; null = legacy (provenance unknown)" }
    DomainStat:
      type: object
      required: [id, domain, domainType, retrievals, deltaPct, isOwned]
      properties:
        id: { type: integer }
        domain: { type: string }
        domainType: { type: string }
        retrievals: { type: integer }
        deltaPct: { type: number }
        isOwned: { type: boolean }
        firstSeenAt: { type: ["string", "null"] }
    DomainTypeDistribution:
      type: object
      required: [totalRetrievals, types]
      properties:
        totalRetrievals: { type: integer }
        types:
          type: array
          items:
            $ref: "#/components/schemas/DomainTypeShare"
    DomainTypeShare:
      type: object
      required: [domainType, retrievals, sharePct]
      properties:
        domainType: { type: string }
        retrievals: { type: integer }
        sharePct: { type: number }
    UrlStat:
      type: object
      required: [id, url, domain, domainType, retrievals]
      properties:
        id: { type: integer }
        url: { type: string }
        domain: { type: string }
        domainType: { type: string }
        retrievals: { type: integer }
        lastSeenAt: { type: ["string", "null"] }
...